refactor(studio): decompose the SQL editor panel
SqlEditorPanel was the last component over react-doctor's 300-line limit, at
714. It is now 193, split along the seams that were already there — three
hooks and four components, no behaviour change:

- `useSqlEditorTabs` — which tabs are open, which is active, each one's
  ephemeral output, and the right-click menu's bulk closes with their
  unsaved-work confirm.
- `useSqlLibrary` — the persisted PRIVATE query list, the run history, and the
  sidebar's search. Reaches the editor through two callbacks, so it never
  needs the tab model.
- `useSqlEditorSurface` — completion, caret tracking, the keyboard map,
  gutter/overlay scroll sync, and diagnostic reveal: everything about the one
  textarea and where its caret is.
- `SqlTabStrip`, `SqlEditorPane`, `SqlResultsPane`, and `SqlResultTable`.

The strip and the two panes take their models whole rather than 18-plus loose
props each: every field is rendered in exactly one of them, so a flattened
signature would restate the hook's shape without decoupling anything.
`onSelect` stays a separate prop because switching tabs also dismisses the
completion popover — the panel's concern, not the strip's.

Verified the extraction is structural: the set of `data-testid`s the panel
renders is byte-identical before and after. 930 studio tests unchanged.

Closes the last `no-giant-component` suppression in the package.
